EmployeeTimeTracking/Validators/DateGreaterThanAttribute.cs
using System.ComponentModel.DataAnnotations;
using Xunit.Sdk;

public class DateGreaterThanAttribute : ValidationAttribute
{
    private readonly string _comparisonProperty;

    public DateGreaterThanAttribute(string comparisonProperty)
    {
        _comparisonProperty = comparisonProperty;
    }

    protected override ValidationResult IsValid(object value, ValidationContext validationContext)
    {
        var currentValue = (DateTime?)value;

        var property = validationContext.ObjectType.GetProperty(_comparisonProperty);

        if (property == null)
        {
            return new ValidationResult($"Unknown property: {_comparisonProperty}");
        }

        var comparisonValue = (DateTime?)property.GetValue(validationContext.ObjectInstance);

        if (currentValue.HasValue && comparisonValue.HasValue && currentValue <= comparisonValue)
        {
            var errorMessage = ErrorMessage ?? $"The {validationContext.DisplayName} must be greater than {_comparisonProperty}.";
            return new ValidationResult(errorMessage);
        }

        return ValidationResult.Success;
    }
}
